+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Finnish translation for "PRECIP", "UPDATE_INFO_MULTIPLE" and "UPDATE_INFO_SINGLE".
 - fix upgrade script download of dumpactivemodule.js too early, cause fatal error
+- fix upgrade script, missing $ on variable name for temp git stash
 
 ### Fixed
 - Force declaration of public ip adress in config file (ISSUE #1852)